Remove code coverage in CI
Reason: Due to the outdated configuration of the Coveralls tool, we are removing code coverage from the project CI. We decided that code coverage can be addressed in a separate PR in the future.
Implementation: Code coverage has been removed from the CI pipeline and README.md.
Side effects: The support for code coverage should be restored in the future.